refactor(config): move registry framework to utils, rename compat->pycompat
Addresses PR #159 review feedback:
- Move the ConfigRegistry/plugin-loading framework out of the user-facing
  config_values/ preset dir into utils/config_registry.py (core + inference).
- Rename compat.py -> pycompat.py to tie the name to Python-version shims.

Mechanical import updates across all callers; no behavior change. The mirrored
core/inference registry NOTE and the entry_points fallback (requires-python
>=3.8) are preserved.
