+++ src/holosoma/holosoma/config_values/command.py
@@ -3,11 +3,11 @@
 from holosoma.config_types.command import CommandManagerCfg
 from holosoma.config_values.loco.g1.command import g1_29dof_command
 from holosoma.config_values.loco.t1.command import t1_29dof_command
-from holosoma.config_values.registry import ConfigRegistry
 from holosoma.config_values.wbt.g1.command import (
     g1_29dof_wbt_command,
     g1_29dof_wbt_command_w_object,
 )
+from holosoma.utils.config_registry import ConfigRegistry
 
 COMMAND_REGISTRY = ConfigRegistry(CommandManagerCfg, group="holosoma.config.command")
 
@@ -17,7 +17,7 @@ COMMAND_REGISTRY.add("g1_29dof", g1_29dof_command)
 COMMAND_REGISTRY.add("g1_29dof_wbt", g1_29dof_wbt_command)
 COMMAND_REGISTRY.add("g1_29dof_wbt_w_object", g1_29dof_wbt_command_w_object)
 
-from holosoma.config_values.registry import (  # noqa: E402
+from holosoma.utils.config_registry import (  # noqa: E402
     deprecated_defaults_alias as _deprecated_defaults_alias,
 )
