+++ src/holosoma/holosoma/config_values/curriculum.py
@@ -3,8 +3,8 @@
 from holosoma.config_types.curriculum import CurriculumManagerCfg
 from holosoma.config_values.loco.g1.curriculum import g1_29dof_curriculum, g1_29dof_curriculum_fast_sac
 from holosoma.config_values.loco.t1.curriculum import t1_29dof_curriculum, t1_29dof_curriculum_fast_sac
-from holosoma.config_values.registry import ConfigRegistry
 from holosoma.config_values.wbt.g1.curriculum import g1_29dof_wbt_curriculum
+from holosoma.utils.config_registry import ConfigRegistry
 
 CURRICULUM_REGISTRY = ConfigRegistry(CurriculumManagerCfg, group="holosoma.config.curriculum")
 
@@ -15,7 +15,7 @@ CURRICULUM_REGISTRY.add("t1_29dof_fast_sac", t1_29dof_curriculum_fast_sac)
 CURRICULUM_REGISTRY.add("g1_29dof_fast_sac", g1_29dof_curriculum_fast_sac)
 CURRICULUM_REGISTRY.add("g1_29dof_wbt_curriculum", g1_29dof_wbt_curriculum)
 
-from holosoma.config_values.registry import (  # noqa: E402
+from holosoma.utils.config_registry import (  # noqa: E402
     deprecated_defaults_alias as _deprecated_defaults_alias,
 )
