+++ src/holosoma/holosoma/config_values/experiment.py
@@ -4,13 +4,13 @@ from typing_extensions import Annotated
 from holosoma.config_types.experiment import ExperimentConfig
 from holosoma.config_values.loco.g1.experiment import g1_29dof, g1_29dof_fast_sac
 from holosoma.config_values.loco.t1.experiment import t1_29dof, t1_29dof_fast_sac
-from holosoma.config_values.registry import ConfigRegistry
 from holosoma.config_values.wbt.g1.experiment import (
     g1_29dof_wbt,
     g1_29dof_wbt_fast_sac,
     g1_29dof_wbt_fast_sac_w_object,
     g1_29dof_wbt_w_object,
 )
+from holosoma.utils.config_registry import ConfigRegistry
 
 EXPERIMENT_REGISTRY = ConfigRegistry(ExperimentConfig, group="holosoma.config.experiment")
 
@@ -36,7 +36,7 @@ def get_annotated_experiment_config() -> type:
     ]
 
 
-from holosoma.config_values.registry import (  # noqa: E402
+from holosoma.utils.config_registry import (  # noqa: E402
     deprecated_defaults_alias as _deprecated_defaults_alias,
 )
