+++ src/holosoma/holosoma/config_values/logger.py
@@ -1,5 +1,5 @@
 from holosoma.config_types.logger import DisabledLoggerConfig, WandbLoggerConfig
-from holosoma.config_values.registry import ConfigRegistry
+from holosoma.utils.config_registry import ConfigRegistry
 
 LOGGER_REGISTRY = ConfigRegistry((DisabledLoggerConfig, WandbLoggerConfig), group="holosoma.config.logger")
 
@@ -7,7 +7,7 @@ disabled = LOGGER_REGISTRY.add("disabled", DisabledLoggerConfig())
 wandb = LOGGER_REGISTRY.add("wandb", WandbLoggerConfig(mode="online"))
 wandb_offline = LOGGER_REGISTRY.add("wandb_offline", WandbLoggerConfig(mode="offline"))
 
-from holosoma.config_values.registry import (  # noqa: E402
+from holosoma.utils.config_registry import (  # noqa: E402
     deprecated_defaults_alias as _deprecated_defaults_alias,
 )
