+++ src/holosoma/holosoma/config_values/scene.py
@@ -17,8 +17,8 @@ from holosoma.config_types.scene import (
     SceneConfig,
     SceneFileConfig,
 )
-from holosoma.config_values.registry import ConfigRegistry, deprecated_defaults_alias
 from holosoma.config_values.wbt.g1.scene import g1_29dof_wbt_object_scene
+from holosoma.utils.config_registry import ConfigRegistry, deprecated_defaults_alias
 
 SCENE_REGISTRY = ConfigRegistry(SceneConfig, group="holosoma.config.scene")
