+++ src/holosoma/holosoma/config_values/termination.py
@@ -3,8 +3,8 @@
 from holosoma.config_types.termination import TerminationManagerCfg
 from holosoma.config_values.loco.g1.termination import g1_29dof_termination
 from holosoma.config_values.loco.t1.termination import t1_29dof_termination
-from holosoma.config_values.registry import ConfigRegistry
 from holosoma.config_values.wbt.g1.termination import g1_29dof_wbt_termination
+from holosoma.utils.config_registry import ConfigRegistry
 
 TERMINATION_REGISTRY = ConfigRegistry(TerminationManagerCfg, group="holosoma.config.termination")
 
@@ -13,7 +13,7 @@ TERMINATION_REGISTRY.add("t1_29dof", t1_29dof_termination)
 TERMINATION_REGISTRY.add("g1_29dof", g1_29dof_termination)
 TERMINATION_REGISTRY.add("g1_29dof_wbt", g1_29dof_wbt_termination)
 
-from holosoma.config_values.registry import (  # noqa: E402
+from holosoma.utils.config_registry import (  # noqa: E402
     deprecated_defaults_alias as _deprecated_defaults_alias,
 )
