+++ src/holosoma_inference/holosoma_inference/config/config_values/observation.py
@@ -7,7 +7,7 @@ robot types and tasks, converted from the original YAML configurations.
 from __future__ import annotations
 
 from holosoma_inference.config.config_types.observation import ObservationConfig
-from holosoma_inference.config.registry import (
+from holosoma_inference.utils.config_registry import (
     ConfigRegistry,
     deprecated_defaults_alias,
     deprecated_get_defaults,
