+++ src/holosoma_inference/holosoma_inference_service/holosoma_service/holosoma_service/policy_control/service_node.py
@@ -45,10 +45,10 @@ from rclpy.qos import QoSProfile, ReliabilityPolicy
 from sensor_msgs.msg import JointState
 
 from holosoma_inference.config.config_values.inference import get_annotated_inference_config
-from holosoma_inference.config.registry import parse_config
 from holosoma_inference.config.utils import TYRO_CONFIG
 from holosoma_inference.policies.dual_mode import DualModePolicy, _select_policy_class
 from holosoma_inference.sensors.base import Sensor
+from holosoma_inference.utils.config_registry import parse_config
 from holosoma_service.policy_control.injected_inputs import (
     CMD_VEL_TOPIC,
     STATE_INPUT_TOPIC,
